# Doc/swagger_subastar.yaml
openapi: 3.0.3
info:
  title: SubastAR API
  description: >
    API REST de SubastAR — app de subastas dinámicas ascendentes (TPO DA1, 1C2026).
    Cubre el dominio descripto en TPO_DAI_1C2026.pdf: registro en dos etapas con
    foto de DNI, gestión de medios de pago, subastas con categoría y moneda,
    pujas con reglas de mínimo/máximo, consignaciones con inspección y póliza
    de seguro, notificaciones, métricas y ventas.
  version: 2.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local
  - url: https://subastar.example.com/api/v1
    description: Producción (entrega 3)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Bid:
      type: object
      properties:
        id: { type: string }
        subastaId: { type: string }
        piezaId: { type: string }
        usuarioId: { type: string }
        usuarioNombre: { type: string }
        monto: { type: number, format: double }
        moneda: { type: string, enum: [ARS, USD] }
        timestamp: { type: string, format: date-time }
        ganadora: { type: boolean }
    BidRequest:
      type: object
      required: [piezaId, monto]
      properties:
        piezaId: { type: string }
        monto: { type: number, format: double }
        medioPagoId: { type: string }
    BidPage:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Bid' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
...
security:
  - bearerAuth: []
paths:
...
  /auctions/{auctionId}/bids:
    get:
      tags: [Bids]
      summary: Historial de pujas de la subasta
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Pujas paginadas.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BidPage' }
    post:
      tags: [Bids]
      summary: Realizar una puja
      description: |
        Reglas (PDF SubastAR):
        - `monto >= mejorOferta + 0.01 * precioBase`
        - `monto <= ultimaOferta + 0.20 * precioBase` (excepto en subastas ORO/PLATINO)
        - El usuario debe tener al menos un medio de pago verificado.
        - Una sola puja por usuario en vuelo (locking en cliente STOMP hasta confirmación).
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BidRequest' }
      responses:
        '201':
          description: Puja registrada (también se broadcast por WebSocket).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bid' }
        '400': { description: Monto fuera de rango (min/max). }
        '403': { description: Sin medio de pago verificado o categoría insuficiente. }
        '409': { description: Hay una puja del usuario aún sin confirmar. }

  # ─── PAYMENT METHODS ───────────────────────────────────────────────
...